% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHadUK.R
\name{getHadUK}
\alias{getHadUK}
\title{Get HadUK weather data}
\usage{
getHadUK(
  username = "roconnellbooth",
  password = "xx",
  version = "v1.3.0.ceda",
  subversion = "v20240514",
  var = "rainfall",
  spatial_resolution = "1km",
  temporal_resolution = "day",
  scratch = "temp",
  northing_col = "northing",
  easting_col = "easting",
  sample_id_col = "sample",
  date_col = "date",
  data = data.frame(northing = 572100, easting = 435900, sample = "X1", date =
    as.Date("2015-10-23"))
)
}
\arguments{
\item{username}{CEDA username}

\item{password}{CEDA password}

\item{version}{dataset version}

\item{subversion}{dataset subversion}

\item{var}{desired weather variable. "rainfall" for rainfall volumn, "tasmax" for max temp, "tasmin" for min temp}

\item{spatial_resolution}{desired spatial_resolution}

\item{temporal_resolution}{desired temporal_resolution}

\item{scratch}{scratch directory to store intermediate .nc files; use "temp" to set to temporary direcory (recommended)}

\item{northing_col}{column containing British National Grid northing (EPSG:27700)}

\item{easting_col}{column containing British National Grid easting (EPSG:27700)}

\item{sample_id_col}{unique identifier for each sample (location, time) for which weather data is required}

\item{date_col}{date column in as.Date() format}

\item{data}{input data}
}
\value{
weather data for specified dates and locations
}
\description{
Get HadUK weather data
}
\examples{
getHadUK(
 username ="roconnellbooth", 
 password = "B33n50nt05t", 
 version = "v1.3.0.ceda", 
 subversion= "v20240514",
 var = "rainfall",  
 spatial_resolution = "1km",
 temporal_resolution = "day",
 scratch = "temp",
 northing_col = "northing",
 easting_col = "easting",
 sample_id_col = "sample",
 date_col = "date",
 data = data.frame(northing = 572100, easting = 435900, sample = "X1", date = as.Date("2015-10-23"))
)
}
